Ability to switch between scenes from scene scope implemented
currentScene is passed into the scenes and stored as a pointer to a
pointer in scene.h. each scene is also given references to the other
scenes via a initializePointers method. For example, this allows the
search scene to change the scene with *currentScene = museumPtr; Control
flow can be determined dynamically inside of the classes.